+++ zkevm-circuits/src/keccak_circuit.rs
@@ -3,5 +3,9 @@
 pub mod keccak_packed_multi;
 /// Util
 pub mod util;
+mod cell_manager;
+#[cfg(test)]
+mod test;
+mod param;
 
 pub use keccak_packed_multi::KeccakCircuitConfig as KeccakConfig;
