+++ zkevm-circuits/src/keccak_circuit/keccak_packed_multi.rs
@@ -23,12 +23,11 @@ use halo2_proofs::{
 };
 use log::{debug, info};
 use std::{env::var, marker::PhantomData, vec};
-
-const MAX_DEGREE: usize = 3;
-const ABSORB_LOOKUP_RANGE: usize = 3;
-const THETA_C_LOOKUP_RANGE: usize = 6;
-const RHO_PI_LOOKUP_RANGE: usize = 4;
-const CHI_BASE_LOOKUP_RANGE: usize = 5;
+use crate::keccak_circuit::{
+    cell_manager::*,
+    param::*,
+    util::*,
+};
 
 pub(crate) fn get_num_rows_per_round() -> usize {
     var("KECCAK_ROWS")
@@ -119,211 +118,6 @@ impl<F: FieldExt> KeccakRegion<F> {
             row.push(F::zero());
         }
         row[column] = value;
-    }
-}
-
-#[derive(Clone, Debug)]
-pub(crate) struct Cell<F> {
-    expression: Expression<F>,
-    column_expression: Expression<F>,
-    column: Option<Column<Advice>>,
-    column_idx: usize,
-    rotation: i32,
-}
-
-impl<F: FieldExt> Cell<F> {
-    pub(crate) fn new(
-        meta: &mut VirtualCells<F>,
-        column: Column<Advice>,
-        column_idx: usize,
-        rotation: i32,
-    ) -> Self {
-        Self {
-            expression: meta.query_advice(column, Rotation(rotation)),
-            column_expression: meta.query_advice(column, Rotation::cur()),
-            column: Some(column),
-            column_idx,
-            rotation,
-        }
-    }
-
-    pub(crate) fn new_value(column_idx: usize, rotation: i32) -> Self {
-        Self {
-            expression: 0.expr(),
-            column_expression: 0.expr(),
-            column: None,
-            column_idx,
-            rotation,
-        }
-    }
-
-    pub(crate) fn at_offset(&self, meta: &mut ConstraintSystem<F>, offset: i32) -> Self {
-        let mut expression = 0.expr();
-        meta.create_gate("Query cell", |meta| {
-            expression = meta.query_advice(self.column.unwrap(), Rotation(self.rotation + offset));
-            vec![0.expr()]
-        });
-
-        Self {
-            expression,
-            column_expression: self.column_expression.clone(),
-            column: self.column,
-            column_idx: self.column_idx,
-            rotation: self.rotation + offset,
-        }
-    }
-
-    pub(crate) fn assign(&self, region: &mut KeccakRegion<F>, offset: i32, value: F) {
-        region.assign(
-            self.column_idx,
-            ((offset as i32) + self.rotation) as usize,
-            value,
-        );
-    }
-
-    pub(crate) fn assign_value(&self, region: &mut KeccakRegion<F>, offset: i32, value: Value<F>) {
-        // This is really ugly. But since there's no way to easily adapt the CellManager
-        // API customized for this impl specifically, for now I'm opening the
-        // value and extracting it. Once https://github.com/privacy-scaling-explorations/zkevm-circuits/issues/933 is resolved,
-        // this shouldn't be needed.
-        let value_f = extract_field(value);
-
-        region.assign(
-            self.column_idx,
-            ((offset as i32) + self.rotation) as usize,
-            value_f,
-        );
-    }
-}
-
-impl<F: FieldExt> Expr<F> for Cell<F> {
-    fn expr(&self) -> Expression<F> {
-        self.expression.clone()
-    }
-}
-
-impl<F: FieldExt> Expr<F> for &Cell<F> {
-    fn expr(&self) -> Expression<F> {
-        self.expression.clone()
-    }
-}
-
-/// CellColumn
-#[derive(Clone, Debug)]
-pub(crate) struct CellColumn<F> {
-    advice: Column<Advice>,
-    expr: Expression<F>,
-}
-
-/// CellManager
-#[derive(Clone, Debug)]
-pub(crate) struct CellManager<F> {
-    height: usize,
-    columns: Vec<CellColumn<F>>,
-    rows: Vec<usize>,
-    num_unused_cells: usize,
-}
-
-impl<F: FieldExt> CellManager<F> {
-    pub(crate) fn new(height: usize) -> Self {
-        Self {
-            height,
-            columns: Vec::new(),
-            rows: vec![0; height],
-            num_unused_cells: 0,
-        }
-    }
-
-    pub(crate) fn query_cell(&mut self, meta: &mut ConstraintSystem<F>) -> Cell<F> {
-        let (row_idx, column_idx) = self.get_position();
-        self.query_cell_at_pos(meta, row_idx as i32, column_idx)
-    }
-
-    pub(crate) fn query_cell_at_row(
-        &mut self,
-        meta: &mut ConstraintSystem<F>,
-        row_idx: i32,
-    ) -> Cell<F> {
-        let column_idx = self.rows[row_idx as usize];
-        self.rows[row_idx as usize] += 1;
-        self.query_cell_at_pos(meta, row_idx, column_idx)
-    }
-
-    pub(crate) fn query_cell_at_pos(
-        &mut self,
-        meta: &mut ConstraintSystem<F>,
-        row_idx: i32,
-        column_idx: usize,
-    ) -> Cell<F> {
-        let column = if column_idx < self.columns.len() {
-            self.columns[column_idx].advice
-        } else {
-            let advice = meta.advice_column();
-            let mut expr = 0.expr();
-            meta.create_gate("Query column", |meta| {
-                expr = meta.query_advice(advice, Rotation::cur());
-                vec![0.expr()]
-            });
-            self.columns.push(CellColumn { advice, expr });
-            advice
-        };
-
-        let mut cells = Vec::new();
-        meta.create_gate("Query cell", |meta| {
-            cells.push(Cell::new(meta, column, column_idx, row_idx));
-            vec![0.expr()]
-        });
-        cells[0].clone()
-    }
-
-    pub(crate) fn query_cell_value(&mut self) -> Cell<F> {
-        let (row_idx, column_idx) = self.get_position();
-        self.query_cell_value_at_pos(row_idx as i32, column_idx)
-    }
-
-    pub(crate) fn query_cell_value_at_row(&mut self, row_idx: i32) -> Cell<F> {
-        let column_idx = self.rows[row_idx as usize];
-        self.rows[row_idx as usize] += 1;
-        self.query_cell_value_at_pos(row_idx, column_idx)
-    }
-
-    pub(crate) fn query_cell_value_at_pos(&mut self, row_idx: i32, column_idx: usize) -> Cell<F> {
-        Cell::new_value(column_idx, row_idx)
-    }
-
-    fn get_position(&mut self) -> (usize, usize) {
-        let mut best_row_idx = 0usize;
-        let mut best_row_pos = 100000usize;
-        for (row_idx, row) in self.rows.iter().enumerate() {
-            if *row < best_row_pos {
-                best_row_pos = *row;
-                best_row_idx = row_idx;
-            }
-        }
-        self.rows[best_row_idx] += 1;
-        (best_row_idx, best_row_pos)
-    }
-
-    pub(crate) fn get_width(&self) -> usize {
-        self.rows.iter().cloned().max().unwrap()
-    }
-
-    pub(crate) fn start_region(&mut self) -> usize {
-        // Make sure all rows start at the same column
-        let width = self.get_width();
-        for row in self.rows.iter_mut() {
-            self.num_unused_cells += width - *row;
-            *row = width;
-        }
-        width
-    }
-
-    pub(crate) fn columns(&self) -> &[CellColumn<F>] {
-        &self.columns
-    }
-
-    pub(crate) fn get_num_unused_cells(&self) -> usize {
-        self.num_unused_cells
     }
 }
 
@@ -603,7 +397,7 @@ mod split_uniform {
                 // Make sure the parts combined equal the value in the uniform output
                 let expr = part_a.expr()
                     + part_b.expr()
-                        * F::from((BIT_SIZE as u32).pow(word_part.bits.len() as u32) as u64);
+                    * F::from((BIT_SIZE as u32).pow(word_part.bits.len() as u32) as u64);
                 cb.require_equal("rot part", expr, output_cells[counter].expr());
 
                 // Input needs the two parts because it needs to be able to undo the rotation
@@ -1546,10 +1340,10 @@ impl<F: Field> SubCircuitConfig<F> for KeccakCircuitConfig<F> {
                     length.clone(),
                     length_prev.clone() * not::expr(start_new_hash_prev.expr())
                         + sum::expr(
-                            is_paddings
-                                .iter()
-                                .map(|is_padding| not::expr(is_padding.expr())),
-                        ),
+                        is_paddings
+                            .iter()
+                            .map(|is_padding| not::expr(is_padding.expr())),
+                    ),
                 );
                 // Use intermediate cells to keep the degree low
                 let mut new_data_rlc =
@@ -1557,7 +1351,7 @@ impl<F: Field> SubCircuitConfig<F> for KeccakCircuitConfig<F> {
                 cb.require_equal("initial data rlc", data_rlcs[0].expr(), new_data_rlc);
                 new_data_rlc = data_rlcs[0].expr();
                 for (idx, (byte, is_padding)) in
-                    input_bytes.iter().zip(is_paddings.iter()).enumerate()
+                input_bytes.iter().zip(is_paddings.iter()).enumerate()
                 {
                     new_data_rlc = select::expr(
                         is_padding.expr(),
@@ -2129,37 +1923,3 @@ fn multi_keccak<F: Field>(
     Ok(rows)
 }
 
-#[cfg(test)]
-mod tests {
-    use super::*;
-    use halo2_proofs::{dev::MockProver, halo2curves::bn256::Fr};
-    use log::error;
-
-    fn verify<F: Field>(k: u32, inputs: Vec<Vec<u8>>, success: bool) {
-        let circuit = KeccakCircuit::new(Some(2usize.pow(k)), inputs);
-
-        let prover = MockProver::<F>::run(k, &circuit, vec![]).unwrap();
-        let verify_result = prover.verify();
-        if verify_result.is_ok() != success {
-            if let Some(errors) = verify_result.err() {
-                for error in errors.iter() {
-                    error!("{}", error);
-                }
-            }
-            panic!();
-        }
-    }
-
-    #[test]
-    fn packed_multi_keccak_simple() {
-        let k = 11;
-        let inputs = vec![
-            vec![],
-            (0u8..1).collect::<Vec<_>>(),
-            (0u8..135).collect::<Vec<_>>(),
-            (0u8..136).collect::<Vec<_>>(),
-            (0u8..200).collect::<Vec<_>>(),
-        ];
-        verify::<Fr>(k, inputs, true);
-    }
-}
